google-pubsub-oneoff/requirements.txt
docopt==0.6.2
google-cloud-pubsub==1.6.1
